third_party/blink/web_tests/platform/fuchsia/external/wpt/css/css-properties-values-api/var-reference-registered-properties-expected.txt

This is a testharness.js-based test.
[FAIL] Invalid values for registered properties are serialized as the empty string
  assert_equals: expected "" but got "0px"
Harness: the test ran to completion.
